==> Source/1 - Tab_Acessos.ipynb <==
# Tabela Acessos

## Obtenção e carregamento de dados  

#importação de bibliotecas
from sqlalchemy import create_engine
import pandas as pd

# Conexão conexão com o banco de dados
engine = create_engine('mysql+pymysql://evasao:evasao2023@localhost:3306/evasao_ead')

# Lendo dados da tabela acessos
df_acessos = pd.read_sql_query("SELECT * FROM acessos", engine)

# Impressão das 5 primeiras linhas do dataframe de acessos
df_acessos.head()

## Informações sobre o dataset

#Exibir informações sobre o dataframe
df_acessos.info()

#Exibir estatísticas descritivas do dataframe
df_acessos.describe()

## Limpeza e Tratatamento de Dados

### Índice

#Definir a coluna cd_acesso como índice
df_acessos.set_index('cd_acesso', inplace=True)

### Data de saída (Logout)

Remover a coluna dt_saida pois não será utilizada na análise visto que não possui informações relevantes para a análise.

#Remover coluna dt_saida
df_acessos.drop('dt_saida', axis=1, inplace=True)
df_acessos.head()

### Data de entrada (Login)

#Formata a data de entrada ('YYYY-MM-DD')
df_acessos['dt_entrada'] = pd.to_datetime(df_acessos['dt_entrada']).dt.normalize()
#Altera o nome da coluna dt_entrada para dt_acesso
df_acessos.rename(columns={'dt_entrada': 'dt_acesso'}, inplace=True)
df_acessos.head()

### Código pessoa (Aluno)

#Altera o nome da coluna cd_pessoa para cd_aluno
df_acessos.rename(columns={'cd_pessoa': 'cd_aluno'}, inplace=True)
df_acessos.head()

## Verificação e Exportação de Dados

### Verificação

#Exibir informações sobre o dataframe
df_acessos.info()

#Exibir estatísticas descritivas do dataframe
df_acessos.describe()

### Exportação

### Exporta o dataframe para um arquivo CSV
df_acessos.to_csv('../Data/tab_Acessos.csv')
