==> Source/3 - Tab_Cursos_Oferta.ipynb <==
# Tabela Cursos Oferta

## Obtenção e carregamento de dados  

#importação de bibliotecas
from sqlalchemy import create_engine
import pandas as pd

# Conexão conexão com o banco de dados
engine = create_engine('mysql+pymysql://evasao:evasao2023@localhost:3306/evasao_ead')

# Lendo dados da tabela cursos_oferta
df_cursos_oferta = pd.read_sql_query("SELECT * FROM cursos_oferta", engine)

# Impressão das 5 primeiras linhas do dataframe de cursos_oferta
df_cursos_oferta.head()

## Informações sobre o dataset

#Exibir informações sobre o dataframe
df_cursos_oferta.info()

#Exibir estatísticas descritivas do dataframe
df_cursos_oferta.describe()

## Limpeza e Tratamento de dados

### Semestre de inicio (Inaurugação do curso)

#Renomeia a coluna primeiro_semestre para dt_inauguracao
df_cursos_oferta.rename(columns={'primeiro_semestre':'dt_inauguracao'}, inplace=True)


df_cursos_oferta.head()

### Quantidade de semestres (Quantidade de semestres de existência do curso)

#Renomeia a coluna qtd_semestres para semestres_existencia
df_cursos_oferta.rename(columns={'qtd_semestres':'semestres_existencia'}, inplace=True)

df_cursos_oferta.head()

### Semestres duração (Duração do curso em semestres)

#Renomeia a coluna duracao_semestres para qtd_semestres_curso
df_cursos_oferta.rename(columns={'duracao_semestres':'qtd_semestres_curso'}, inplace=True)

df_cursos_oferta.head()

### Curso

#Renomeia a coluna ds_curso para curso
df_cursos_oferta.rename(columns={'ds_curso':'curso'}, inplace=True)

#Remove a coluna ds_apelido
df_cursos_oferta.drop(columns=['ds_apelido'], inplace=True)

df_cursos_oferta.head()

### Ordenação de dados

#Ordena as colunas do dataframe na ordem desejada
df_cursos_oferta = df_cursos_oferta[['cd_curso', 'curso', 'dt_inauguracao', 'semestres_existencia', 'qtd_semestres_curso']]

df_cursos_oferta.head()

## Verificação e Exportação

### Verificação

#Exibir informações sobre o dataframe
df_cursos_oferta.info()

#Exibir estatísticas descritivas do dataframe
df_cursos_oferta.describe()

### Exportação

### Exporta o dataframe para um arquivo CSV
df_cursos_oferta.to_csv('../Data/tab_Cursos_Oferta.csv', index=False)
