==> Source/5 - Tab_Motivo.ipynb <==
# Tabela Motivo

## Obtenção e carregamento de dados  

#importação de bibliotecas
from sqlalchemy import create_engine
import pandas as pd

# Conexão conexão com o banco de dados
engine = create_engine('mysql+pymysql://evasao:evasao2023@localhost:3306/evasao_ead')

# Lendo dados da tabela motivo
df_motivo = pd.read_sql_query("SELECT * FROM motivo", engine)

# Impressão das 5 primeiras linhas do dataframe de motivo
df_motivo.head()

## Informações sobre o dataset

#Exibir informações sobre o dataframe
df_motivo.info()

#Exibir estatísticas descritivas do dataframe
df_motivo.describe()

## Limpeza e Tratatamento de Dados

### Data de registro (Desligamento)

#Formata a data de entrada ('YYYY-MM-DD')
df_motivo['dt_registro'] = pd.to_datetime(df_motivo['dt_registro']).dt.normalize()

#Renomeia a coluna dt_registro para dt_saida
df_motivo.rename(columns={'dt_registro': 'dt_saida'}, inplace=True)

df_motivo.head()

### Código Pessoa (Aluno)

#Renomeia a coluna cd_pessoa para cd_aluno
df_motivo.rename(columns={'cd_pessoa': 'cd_aluno'}, inplace=True)

df_motivo.head()

### Detalhes do Motivo

Remover a coluna tx_motivo, pois ela é descritiva e como não faremos nenhuma análise deste tipo, a informação não será utilizada na análise.

#Remover coluna tx_motivo
df_motivo.drop('tx_motivo', axis=1, inplace=True)
df_motivo.head()

### Ordenação dos dados

#Ordena as colunas do dataframe na ordem desejada
df_motivo = df_motivo[['cd_aluno','cd_curso', 'dt_saida' ,'ds_motivo' ]]

df_motivo.head()

## Verificação e Exportação

### Verificação

#Exibir informações sobre o dataframe
df_motivo.info()

#Exibir estatísticas descritivas do dataframe
df_motivo.describe()

### Exportação

### Exporta o dataframe para um arquivo CSV 
df_motivo.to_csv('../Data/tab_Motivo.csv', index=False)
