==> Source/6 - Tab_Motivo_Classificacao.ipynb <==
# Tabela Motivos Classificação

## Obtenção e carregamento de dados  

#importação de bibliotecas
from sqlalchemy import create_engine
import pandas as pd

# Conexão conexão com o banco de dados
engine = create_engine('mysql+pymysql://evasao:evasao2023@localhost:3306/evasao_ead')

# Lendo dados da tabela motivo
df_motivo_classificacao = pd.read_sql_query("SELECT * FROM motivos_classificacao", engine)

# Impressão das 5 primeiras linhas do dataframe de motivo_classificacao
df_motivo_classificacao.head()

## Informações sobre o dataset

#Exibir informações sobre o dataframe
df_motivo_classificacao.info()

#Exibir estatísticas descritivas do dataframe
df_motivo_classificacao.describe()

# Exibir linhas com a coluna em branco, para a fase de tratamento
df_motivo_classificacao[df_motivo_classificacao.isna().any(axis=1)]

## Limpeza e Tratamento de dados

### Classificação (motivos)

Criamos 5 classificações que determinam dentre os principais motivos, qual o caráter dele, considerando se está relacionado a:
- Curso ou Instituição
- Questões Pessoais ou de Saúde
- Circunstâncias Externas
- Questões Financeiras
- Problemas Técnicos ou Administrativos

# Defina a função para categorizar os motivos
def categorizar_motivos(motivo):
    if motivo in ['Dificuldade em acompanhar o curso', 'Dúvida sobre escolha do curso',
                  'Falta de Identificação com Metodologia', 'Falta de Identificação com o Curso',
                  'Insatisfação com Análise Curricular', 'Insatisfação', 'Insatisfação com Atendimento Unidade',
                  'Insatisfação com o PIPA', 'Insatisfação com o Professor', 'Insatisfação com Matriz Curricular',
                  'Irá para Curso Presencial', 'Mudança de curso', 'Não se adaptou a modalidade EAD',
                  'Não se adaptou a Plataforma de Estudos (Moodle/Portal)','Insafisfação com Análise Curricular']:
        return 'Curso ou Instituição'
    elif motivo in ['Doença', 'Óbito', 'Gestação/RN', 'Razões Particulares','Razões profissionais','Falta de tempo para conciliar estudos e demais atividades',
                    'Falta de equipamento para estudar']:
        return 'Questões Pessoais ou de Saúde'
    elif motivo in ['Aprovado em Faculdade Federal'	,'Aprovado em Faculdade Federal','Mudança de cidade', 'Mudança de País', 'Viagem']:
        return 'Circunstâncias Externas'
    elif motivo in ['Desemprego', 'Razões financeiras', 'Recebeu bolsa de maior valor em outra IES', 'Não Pagou Matrícula', 'Sem acesso a Internet', 'Não realizou a rematrícula']:
        return 'Questões Financeiras'
    else:
        return 'Problemas Técnicos ou Administrativos'

# Aplicar a função à coluna ds_motivo para criar a coluna categoria
df_motivo_classificacao['ds_classificacao'] = df_motivo_classificacao['ds_motivo'].apply(lambda x: categorizar_motivos(x))

#Renomeia a coluna ds_classificacao para categoria_motivo_cancelamento
df_motivo_classificacao.rename(columns={'ds_classificacao': 'categoria_motivo_cancelamento'}, inplace=True)

# Mostrar o DataFrame resultante
df_motivo_classificacao.head()

df_motivo_classificacao['categoria_motivo_cancelamento'].unique()

## Verificação e Exportação

### Verificação

#Exibir informações sobre o dataframe
df_motivo_classificacao.info()

#Exibir estatísticas descritivas do dataframe
df_motivo_classificacao.describe()

### Exportação

### Exporta o dataframe para um arquivo CSV
df_motivo_classificacao.to_csv('../Data/tab_Motivo_Classificacao.csv', index=False)
